# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the core REST API for Graph Auth.",
        "title": "Graph Auth",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/auth/session/consume-code": {
            "post": {
                "description": "Redeems a valid 6-digit delegation code to provision a new child session node in the Neo4j provenance graph. Requested scopes must be limited to read and fertile and be a subset of the parent session scopes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session Delegation"
                ],
                "summary": "Consume Session Invitation Code",
                "parameters": [
                    {
                        "description": "JSON body containing the 6-digit code, device name, and requested scopes",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ConsumeDelegationCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a brand new persistent session token and active scopes",
                        "schema": {
                            "$ref": "#/definitions/api.ConsumeDelegationCodeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format or requested scopes are not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The code has expired, been used, or is mathematically invalid",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Parent session was revoked, inactive, lacks fertile scope, or requested scopes exceed parent scopes",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/session/generate-code": {
            "post": {
                "description": "Generates a temporary, single-use 6-digit code or URL link from an active, fertile session to invite a new device. Requested scopes are chosen when the code is consumed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session Delegation"
                ],
                "summary": "Generate Session Invitation Code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e where token is a Session.token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the 6-digit code, direct link, and TTL expiry",
                        "schema": {
                            "$ref": "#/definitions/api.GenerateDelegationCodeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized due to missing, invalid, or inactive session",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Parent session lacks the fertile scope required to delegate",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal failure generating crypto code or saving to cache",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/session/invalidate": {
            "post": {
                "description": "Permanently deactivates the authenticated session and every descendant it has spawned. Used on logout — cannot be undone.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session Delegation"
                ],
                "summary": "Invalidate Session Tree",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session and all descendants invalidated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid session",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ConsumeDelegationCodeRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.ConsumeDelegationCodeResponse": {
            "type": "object",
            "properties": {
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "session_token": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "authenticated"
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_request"
                }
            }
        },
        "api.GenerateDelegationCodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                }
            }
        },
...
    }
}